rework on how the doc system is working
Each definition can have the following keys related to docs:
:title, :mdkey, :doc.
If :title is not given, the *first* entry of the keys is used
between surrounding backticks. That means that a simple command
like version will have :title set to `version`
If :mdkey is not given, the *first* entry of the keys is used.
The :mdkey value is used when shipping out to markdown as label
for the title

Output to markdown is formatted as follows:
   "## ~a ## {#~a}~2%~a~2%" :title :mdkey :doc

Note: entries *without* a :doc key will not be handled by the doc system